app/whisper-transcription/requirements.txt
flask==2.2.2
gunicorn==20.1.0
whisper-openai